+++ app/util/__init__.py
@@ -1,1 +1,2 @@
 from app.util.completion_executor import CompletionExecutor
+from app.util.news_comments_crawler import NewsCommentsCrawler
